# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayOn futsal management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courts/{id}:
    get:
      operationId: getCourt
      tags: [courts]
      summary: Get a court
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Court
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Court"
        "404":
          description: Not found

  # ── SEASONS ─────────────────────────────────────────────────────────────────
...
  /admin/facility-split-rules:
    get:
      operationId: listFacilitySplitRules
      tags: [admin]
      summary: List facility split rules (admin)
      parameters:
        - name: venueId
          in: query
          schema:
            type: integer
        - name: latestOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Split rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacilitySplitRule"
    post:
      operationId: createFacilitySplitRule
      tags: [admin]
      summary: Create a facility split rule (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilitySplitRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilitySplitRule"
  /admin/facility-split-rules/{id}:
    get:
      operationId: getFacilitySplitRule
      tags: [admin]
      summary: Get a facility split rule (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Split rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilitySplitRule"
        "404":
          description: Not found
    patch:
      operationId: updateFacilitySplitRule
      tags: [admin]
      summary: Update a facility split rule — creates a new version (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacilitySplitRuleUpdate"
      responses:
        "200":
          description: New version of the split rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilitySplitRule"
        "404":
          description: Not found
    delete:
      operationId: deleteFacilitySplitRule
      tags: [admin]
      summary: Deactivate a facility split rule (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deactivated

  # ── SERVICE FEE CONFIG ────────────────────────────────────────────────────
...
components:
  schemas:
...
    Court:
      type: object
      required: [id, name, type, availableForScheduling]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [full, small_sided]
        description:
          type: ["string", "null"]
        availableForScheduling:
          type: boolean
        maxPlayers:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FacilitySplitRule:
      type: object
      required: [id, name, splitType, version, isLatest, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        venueId:
          type: ["integer", "null"]
        offeringType:
          type: ["string", "null"]
        offeringId:
          type: ["integer", "null"]
        splitType:
          type: string
          enum: [percentage, flat, hybrid]
        facilityPct:
          type: ["number", "null"]
        flatFee:
          type: ["number", "null"]
        flatFeeUnit:
          type: ["string", "null"]
        version:
          type: integer
        isLatest:
          type: boolean
        supersededById:
          type: ["integer", "null"]
        isActive:
          type: boolean
        notes:
          type: ["string", "null"]
        createdByClerkId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FacilitySplitRuleInput:
      type: object
      required: [name, splitType]
      properties:
        name:
          type: string
          minLength: 1
        venueId:
          type: ["integer", "null"]
        offeringType:
          type: ["string", "null"]
        offeringId:
          type: ["integer", "null"]
        splitType:
          type: string
          enum: [percentage, flat, hybrid]
        facilityPct:
          type: ["number", "null"]
        flatFee:
          type: ["number", "null"]
        flatFeeUnit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    FacilitySplitRuleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        venueId:
          type: ["integer", "null"]
        offeringType:
          type: ["string", "null"]
        offeringId:
          type: ["integer", "null"]
        splitType:
          type: string
          enum: [percentage, flat, hybrid]
        facilityPct:
          type: ["number", "null"]
        flatFee:
          type: ["number", "null"]
        flatFeeUnit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Service Fee Config ───────────────────────────────────────────────────
...